backend/requirements.txt
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
makefun==1.15.0
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.5
outcome==1.2.0
packaging==23.0
passlib==1.7.4
pendulum==2.1.2
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.4
pyflakes==3.0.1
PyJWT==2.6.0
pypika-tortoise==0.1.6
pytest==7.2.1
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
redis==4.4.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqladmin==0.8.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlalchemy_database==0.1.0
sqlmodel==0.0.8
sqlmodelx==0.0.4
starlette==0.22.0
tomli==2.0.1
tortoise-orm==0.19.2
typing_extensions==4.4.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
WTForms==3.0.1
zipp==3.11.0
flower
fastapi
gunicorn
pymysql
cryptography
fastapi-users[sqlalchemy]
aiomysql
loguru